Renamed `offenderSearchUtils` module
to `prisonerLocationUtils` as this module contains
functions to help determine the prisoner's location.
